Update move cardFilters into cards
