+++ src/components/cards/cardFilters/AbilitiesFilter.js
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux'
-import { setFilters, fetchCards } from '../../redux/cards/actions'
+import { setFilters, fetchCards } from '../../../redux/cards/actions'
 import { Filter } from './Filter'
 
-function EffectFilterUI(props) {
+function AbilitiesFilterUI(props) {
   const handleOnChange = (filter) => {
     let tempFilters = props.filters
     tempFilters[filter.name] = filter.values
@@ -13,16 +13,15 @@ function EffectFilterUI(props) {
 
   return (
     <Filter
-      name="effect"
+      name="abilities"
       handleOnChange={handleOnChange}
       options={[
-        { label: 'Weather', value: 'weather' },
-        { label: 'Decoy', value: 'decoy' },
-        { label: "Commander's horn", value: "commander's horn" },
-        { label: 'Scorch', value: 'scorch' },
-        { label: 'Summon Avenger', value: 'summon avenger' },
-        { label: 'Mardroeme', value: 'mardroeme' },
-        { label: 'Berserker', value: 'berserker' },
+        { label: 'Hero', value: 'hero' },
+        { label: 'Medic', value: 'medic' },
+        { label: 'Moral boost', value: 'morale boost' },
+        { label: 'Muster', value: 'muster' },
+        { label: 'Spy', value: 'spy' },
+        { label: 'Tight bond', value: 'tight bond' },
       ]}
       reset={props.reset}
     />
@@ -39,7 +38,7 @@ const mapDispatchToProps = (dispatch) => ({
   fetchCards: () => dispatch(fetchCards()),
 })
 
-export const EffectFilter = connect(
+export const AbilitiesFilter = connect(
   mapStateToProps,
   mapDispatchToProps
-)(EffectFilterUI)
+)(AbilitiesFilterUI)
