+++ src/components/cards/cardFilters/EffectFilter.js
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux'
-import { setFilters, fetchCards } from '../../redux/cards/actions'
+import { setFilters, fetchCards } from '../../../redux/cards/actions'
 import { Filter } from './Filter'
 
-function AbilitiesFilterUI(props) {
+function EffectFilterUI(props) {
   const handleOnChange = (filter) => {
     let tempFilters = props.filters
     tempFilters[filter.name] = filter.values
@@ -13,15 +13,16 @@ function AbilitiesFilterUI(props) {
 
   return (
     <Filter
-      name="abilities"
+      name="effect"
       handleOnChange={handleOnChange}
       options={[
-        { label: 'Hero', value: 'hero' },
-        { label: 'Medic', value: 'medic' },
-        { label: 'Moral boost', value: 'morale boost' },
-        { label: 'Muster', value: 'muster' },
-        { label: 'Spy', value: 'spy' },
-        { label: 'Tight bond', value: 'tight bond' },
+        { label: 'Weather', value: 'weather' },
+        { label: 'Decoy', value: 'decoy' },
+        { label: "Commander's horn", value: "commander's horn" },
+        { label: 'Scorch', value: 'scorch' },
+        { label: 'Summon Avenger', value: 'summon avenger' },
+        { label: 'Mardroeme', value: 'mardroeme' },
+        { label: 'Berserker', value: 'berserker' },
       ]}
       reset={props.reset}
     />
@@ -38,7 +39,7 @@ const mapDispatchToProps = (dispatch) => ({
   fetchCards: () => dispatch(fetchCards()),
 })
 
-export const AbilitiesFilter = connect(
+export const EffectFilter = connect(
   mapStateToProps,
   mapDispatchToProps
-)(AbilitiesFilterUI)
+)(EffectFilterUI)
